test: footer
-Check if all elements are rendered.
