+++ client/src/components/layout/Footer.tsx
@@ -1,9 +1,11 @@
 import { NavLink } from "react-router-dom";
+import Logo from "../navigation/Logo";
 
 function Footer() {
   return (
     <footer className="flex flex-col items-center w-full  bg-slate-700 text-white text-center">
       <nav role="footer" className="flex w-3/4 justify-evenly m-5">
+        <Logo setShowMobileMenu={() => {}} />
         <ul>
           <li>
             <NavLink to="/" className="flex p-5">
@@ -20,11 +22,6 @@ function Footer() {
               Games
             </NavLink>
           </li>
-          {/* <li>
-            <NavLink to="/leaderboard" className="flex p-5">
-              Leaderboard
-            </NavLink>
-          </li> */}
         </ul>
         <ul>
           <li>
